[feat]: Implement search feature
Add search feature to sos that enables a user
or an agent to search for tickets. The search
feature uses the elasticsearch search engine.
The file search.py contains search methods which
index tickets. The searchable_mixin module gives the
ticket module the ability to automatically manage an associated
full-text index.