[feat]: update user authentication Additional user authentication features such as changing passwords, email accounts, resetting passwords.
Changing passwords/Updating a known password
Enable the user to update his password as long as he is logged in. If user is logged in, he can input
the old/current password, the new password then confirm the new password. This feature was implemented to accomodate
users who might need to change their passwords constantly

Resetting Passwords
Forgot your password? This sub-feature of the authentication feature takes care of that. No login is required as the
user is locked out of their account. The user will be sent an email with instructions to change his password if
forgotten.

Changing a user's email
A user can change their email address. However before doing so, the user must be logged in and when the
user requests a new email address, the new email address must be verified with a confirmation email where a
user is emailed a token with the to-be email address.